Fixed fold_vector, x_train, and y_train
Think I fixed the components of X_train, y_train, and fold_vector. y_train and x_train are now being created via loops. The fold vector is being created with enough 1-to-K elements to be equal to the amount of rows in the original x_mat and elements in y_vec