Works will entire csv file - csv files up to date
This one works with 20 neighbors and 5 num folds only right now. This is because I'm only taking in 4000/4601 lines of the csv file. The two csv files that are important here are as follows:

prediction_dictionary.csv

Note: this should be plotted as validation error vs number of neighbors (with there being the same number of lines as their are total folds. I.e. if the highest number of folds is 5, there should be 5 lines)
This csv file, as the header shows, holds the following data in the order of columns:
> the number of neighbors used to calculate these
> the number of folds used in the NN algorithm
> the predicted value that the NN algorithm gave based on X_new (the tested dataset)
> the actual value that the NN algorithm should have given based on X_new (aka, the values of the equivalent y vector)
> whether or not the NN algorithm predicted correctly

mean_error.csv

This csv file, as the header shows, holds the following data in the order of columns:
> the number of neighbors
> the mean that was calculated for each zero-one-loss output of the neighbors (note that this is the percent error. To find the percent accuracy, you need to subtract these values from 100)
- i.e., when X_new (aka, the tested dataset) was run through the trained NN algorithm (which was trained with the variables X_train, y_train), it produced a resulting vector of predictions. This vector was compared to the actual vector that corresponded with X_new that had the actual numbers the algorithm was supposed to predict. The zero-one-error of these were calculated (if the prediction was correct, a 0 was given. If not, a 1 was given) and then this is averaged to produce the result of how much error this given function had for the given fold and neighbor values
- when all the values for a single neighbor value were generated, these were all averaged to see the result of how well that neighbor did in total over all of the different folds. Those are the values that are given in this csv